+++ .test_calcipy_project/tests/test_zz_test_project.py
@@ -3,7 +3,6 @@
 from pathlib import Path
 
 import tomli
-
 from test_project import __version__