+++ calcipy/code_tag_collector.py
@@ -1,11 +1,16 @@
 """Collect code tags and output for review in a single location."""
 
 import re
+import subprocess
 from collections import defaultdict
+from functools import lru_cache
+from io import BufferedReader
 from pathlib import Path
-from typing import Dict, List, Optional, Pattern, Sequence
+from typing import Dict, List, Optional, Pattern, Sequence, Tuple
 
 import attr
+import pandas as pd
+import pendulum
 from attrs_strict import type_validator
 from beartype import beartype
 from loguru import logger
@@ -27,6 +32,23 @@ Requires formatting with list of tags: `CODE_TAG_RE.format(tag='|'.join(tag_list
 Commonly, the `tag_list` could be `COMMON_CODE_TAGS`
 
 """
+
+
+@beartype
+def _run_cmd(cmd: str, **kwargs) -> str:
+    """Run command with subprocess and return the output.
+
+    Args:
+        cmd: string command
+        kwargs: any additional keyword arguments to pass to `subprocess.Popen` (typically `cwd`)
+
+    Returns:
+        str: stripped output
+
+    """
+    proc = subprocess.Popen(cmd, shell=True, stdout=subprocess.PIPE, universal_newlines=True, **kwargs)
+    stdout: BufferedReader = proc.stdout  # type: ignore
+    return stdout.read().strip()
 
 
 @attr.s(auto_attribs=True)
@@ -101,6 +123,77 @@ def _search_files(paths_source: Sequence[Path], regex_compiled: Pattern[str]) ->
     return matches
 
 
+@lru_cache
+@beartype
+def _git_info(cwd: Path) -> Tuple[Path, str]:
+    """Collect information about the local git repository.
+
+    Based on snippets from: https://gist.github.com/abackstrom/4034721#gistcomment-3982270
+    and: https://github.com/rscherf/GitLink/blob/e2e7c412630246efc86de4fe71192f15bf11209e/GitLink.py
+
+    Args:
+        cwd: Path to the current working directory (typically file_path.parent)
+
+    Returns:
+        Tuple[Path, str]: (git_dir, repo_url)
+
+    """
+    git_dir = Path(_run_cmd('git rev-parse --show-toplevel', cwd=str(cwd)))
+    clone_uri = _run_cmd('git remote get-url origin', cwd=str(cwd))
+    # Could be ssh or http (with or without .git)
+    # git@github.com:KyleKing/calcipy.git
+    # https://github.com/KyleKing/calcipy.git
+    sub_url = re.findall(r'^.+github.com[:/]([^.]+)(?:\.git)?$', clone_uri)[0]
+    repo_url = f'https://github.com/{sub_url}'
+    return (git_dir, repo_url)
+
+
+@beartype
+def _format_record(base_dir: Path, file_path: Path, comment: _CodeTag) -> Dict[str, str]:
+    """Format each table row for the code tag summary file. Include git permalink.
+
+    Args:
+        base_dir: base path of the project if git directory is not known
+        file_path: path to the file of interest
+        comment: _CodeTag information for the matched tag
+
+    Returns:
+        Dict[str, str]: formatted dictionary with file info
+
+    """
+    cwd = file_path.parent
+    git_dir, repo_url = _git_info(cwd=cwd)
+    blame = _run_cmd(f'git blame {file_path} -L {comment.lineno},{comment.lineno} --porcelain', cwd=cwd)
+    # Set fallbacks if git logic doesn't work
+    source_file = f'{file_path.relative_to(base_dir).as_posix()}:{comment.lineno}'
+    ts = 'N/A'
+    if blame:
+        # Note: line number may be different in older blame (and relative path)
+        revision, old_line_number = blame.split('\n')[0].split(' ')[:2]
+        # If the change has not yet been committed, use the branch name as best guess
+        if all(_c == '0' for _c in revision):
+            revision = _run_cmd('git branch --show-current', cwd=cwd)
+        # Format a nice timestamp of the last edit to the line
+        blame_dict = {
+            line.split(' ')[0]: ' '.join(line.split(' ')[1:])
+            for line in blame.split('\n')
+        }
+        dt = pendulum.from_timestamp(int(blame_dict['committer-time']))
+        tz = blame_dict['committer-tz'][:3] + ':' + blame_dict['committer-tz'][-2:]
+        ts = pendulum.parse(dt.isoformat()[:-6] + tz).format('YYYY-MM-DD')
+        remote_file_path = blame_dict['filename']
+        # PLANNED: Consider making "blame" configurable
+        git_url = f'{repo_url}/blame/{revision}/{remote_file_path}#L{old_line_number}'
+        source_file = f'[{source_file}]({git_url})'
+
+    return {
+        'Type': f'{comment.tag:>7}',
+        'Comment': comment.text,
+        'Last Edit': ts,
+        'Source File': source_file,
+    }
+
+
 @beartype
 def _format_report(
     base_dir: Path, code_tags: List[_Tags], tag_order: List[str],
@@ -117,21 +210,23 @@ def _format_report(
 
     """
     output = ''
+    records = []
     counter: Dict[str, int] = defaultdict(lambda: 0)
     for comments in sorted(code_tags, key=lambda tc: tc.path_source, reverse=False):
-        output += f'- {comments.path_source.relative_to(base_dir).as_posix()}\n'
         for comment in comments.code_tags:
             if comment.tag in tag_order:
-                output += f'    - line {comment.lineno:>3} {comment.tag:>7}: {comment.text}\n'
+                records.append(_format_record(base_dir, comments.path_source, comment))
                 counter[comment.tag] += 1
-        output += '\n'
+    if records:
+        df_tags = pd.DataFrame(records)
+        output += df_tags.to_markdown(index=False)
     logger.debug('counter={counter}', counter=counter)
 
     sorted_counter = {tag: counter[tag] for tag in tag_order if tag in counter}
     logger.debug('sorted_counter={sorted_counter}', sorted_counter=sorted_counter)
     formatted_summary = ', '.join(f'{tag} ({count})' for tag, count in sorted_counter.items())
     if formatted_summary:
-        output += f'Found code tags for {formatted_summary}\n'
+        output += f'\n\nFound code tags for {formatted_summary}\n'
     return output